requirements.txt
python-telegram-bot==20.7
requests==2.31.0
flask==2.3.3
gunicorn>=21.2.0
psutil>=5.9.0
python-dotenv==1.0.0
werkzeug>=3.0.0  # Important for Flask compatibility
pillow>=10.0.0
